* 'dev' of github.com:react-native-kit/react-native-track-player: (39 commits)
  [iOS] Add missing import.
  [iOS] Use React Native utility from RCTConvert to parse provided urls. (#950)
  [ios] resync playback values when updating NowPlayingInfoCenter (#958)
  [ios] deduce togglePlayPause from play and pause capabilities (#957)
  [ios] fix race conditions (close #652) (#956)
  [js] Update hooks (#935)
  [docs] Fixed broken links in ReadMe Documentation (#959)
  [docs] Added a note for custom icons in debug mode #475
  [js] Fixed local artworks not working in updateMetadataForTrack. Closes #944
  [iOS] Reduce swift version to 4.2 for better backwards compatibility with older Xcode versions.
  [docs] Document inability to test lock screen controls on recent iOS simulators. Close #914.
  Clone tracks array in add() to avoid issues with immutable objects. Fixes #930. Fixes #742. Fixes #597.
  [docs] Clarify casting support in the docs. (#926)
  [docs] Add instructions for running react-native-swift during installation.
  [docs] HLS is functioning on iOS (#928)
  Rework docs (#925)
  [android] Fixed position not updating when track changes. Closes #920
  [docs] Update README to include example screenshots. Closes #354
  [example][iOS] Add iOS screenshots.
  [example] Use artwork URLs that will not randomly redirect, so that lock screen image matches image shown in app.
  ...

+++ ios/RNTrackPlayer/RNTrackPlayer.swift
@@ -263,8 +263,12 @@ public class RNTrackPlayer: RCTEventEmitter {
     
     @objc(updateOptions:resolver:rejecter:)
     public func update(options: [String: Any], resolve: RCTPromiseResolveBlock, reject: RCTPromiseRejectBlock) {
-        let capabilitiesStr = options["capabilities"] as? [String]
-        let capabilities = capabilitiesStr?.compactMap { Capability(rawValue: $0) } ?? []
+
+        var capabilitiesStr = options["capabilities"] as? [String] ?? []
+        if (capabilitiesStr.contains("play") && capabilitiesStr.contains("pause")) {
+            capabilitiesStr.append("togglePlayPause");
+        }
+        let capabilities = capabilitiesStr.compactMap { Capability(rawValue: $0) }
         
         let remoteCommands = capabilities.map { capability in
             capability.mapToPlayerCommand(jumpInterval: options["jumpInterval"] as? NSNumber,
@@ -322,19 +326,20 @@ public class RNTrackPlayer: RCTEventEmitter {
     @objc(remove:resolver:rejecter:)
     public func remove(tracks ids: [String], resolve: RCTPromiseResolveBlock, reject: RCTPromiseRejectBlock) {
         print("Removing tracks:", ids)
-        var indexesToRemove: [Int] = []
         
-        for id in ids {
-            if let index = player.items.firstIndex(where: { ($0 as! Track).id == id }) {
-                if index == player.currentIndex { return }
-                indexesToRemove.append(index)
+        // Look through queue for tracks that match one of the provided ids.
+        // Do this in reverse order so that as they are removed, 
+        // it will not affect other indices that will be removed.
+        for (index, element) in player.items.enumerated().reversed() {
+            // skip the current index
+            if index == player.currentIndex { continue }
+            
+            let track = element as! Track
+            if ids.contains(track.id) {
+                try? player.removeItem(at: index)
             }
         }
-        
-        for index in indexesToRemove {
-            try? player.removeItem(at: index)
-        }
-        
+
         resolve(NSNull())
     }
     
@@ -502,13 +507,14 @@ public class RNTrackPlayer: RCTEventEmitter {
                 MediaItemProperty.title(track.title),
                 MediaItemProperty.albumTitle(track.album),
             ])
-            
+            player.updateNowPlayingPlaybackValues();
             track.getArtwork { [weak self] image in
                 if let image = image {
                     let artwork = MPMediaItemArtwork(boundsSize: image.size, requestHandler: { (size) -> UIImage in
                         return image
                     })
                     self?.player.nowPlayingInfoController.set(keyValue: MediaItemProperty.artwork(artwork))
+                    self?.player.updateNowPlayingPlaybackValues();
                 }
             }
         }
